refactor(utils): promote card_data to services/card_data_service
Split utils/card_data.py (408 lines, one CardDataManager class doing six
responsibilities) into a services/card_data_service/ package, mirroring
the gamelog_service / mana_icon_service / image_service precedents:

- schemas.py: CardEntry (with dict-compat shims) and CardIndex
- remote.py: HEAD + GET against MTGJSON AtomicCards (curl_cffi)
- builder.py: pure normalization of raw MTGJSON dicts into index entries
- storage.py: index/meta paths, msgspec decoder, atomic writes
- service.py: CardDataManager orchestrator + load_card_manager factory
- protocol.py: CardDataManagerProto for typed dependency injection

Updates the 18 importers across controllers, repositories, services,
widgets, and tests. test_card_data_refresh.py now monkeypatches
services.card_data_service.remote.requests; test_card_data_aliases.py
calls the module-level build_index directly. ARCHITECTURE.md moves the
node from the Utilities subgraph into the Services subgraph.
