--- utils/card_data.py
@@ -1,407 +1,0 @@
-from __future__ import annotations
-
-import hashlib
-import io
-import json
-import zipfile
-from pathlib import Path
-from typing import Any
-
-import msgspec
-import msgspec.json
-from curl_cffi import requests
-from loguru import logger
-
-from utils.atomic_io import atomic_write_json
-from utils.constants import (
-    ATOMIC_DATA_DOWNLOAD_TIMEOUT_SECONDS,
-    ATOMIC_DATA_HEAD_TIMEOUT_SECONDS,
-    ATOMIC_DATA_URL,
-    CARD_DATA_DIR,
-)
-
-# ---------------------------------------------------------------------------
-# msgspec schemas for the atomic-cards index
-# ---------------------------------------------------------------------------
-
-
-class CardEntry(msgspec.Struct, gc=False):
-    """A single card record as stored in the local atomic-cards index.
-
-    ``gc=False`` disables cyclic-garbage-collection tracking for this type,
-    which measurably reduces parse time when decoding large lists of structs.
-    """
-
-    name: str
-    name_lower: str
-    aliases: list[str]
-    colors: list[str]
-    color_identity: list[str]
-    legalities: dict[str, str]
-    mana_cost: str | None = None
-    mana_value: float | None = None
-    type_line: str | None = None
-    oracle_text: str | None = None
-    power: str | None = None
-    toughness: str | None = None
-    loyalty: str | None = None
-    # Back-face fields populated for double-faced/split/MDFC/adventure cards.
-    back_name: str | None = None
-    back_mana_cost: str | None = None
-    back_type_line: str | None = None
-    back_oracle_text: str | None = None
-    back_power: str | None = None
-    back_toughness: str | None = None
-    back_loyalty: str | None = None
-
-    # ------------------------------------------------------------------
-    # Dict-compatible accessors so that existing callers that treat the
-    # returned value as a dict continue to work without modification.
-    # ------------------------------------------------------------------
-
-    def get(self, key: str, default: Any = None) -> Any:  # noqa: ANN401
-        return getattr(self, key, default)
-
-    def __getitem__(self, key: str) -> Any:  # noqa: ANN401
-        try:
-            return getattr(self, key)
-        except AttributeError:
-            raise KeyError(key) from None
-
-    def __contains__(self, key: object) -> bool:
-        return isinstance(key, str) and hasattr(self, key)
-
-
-class CardIndex(msgspec.Struct):
-    """Root structure of the saved atomic-cards index file."""
-
-    cards: list[CardEntry]
-    cards_by_name: dict[str, CardEntry]
-
-
-# Pre-instantiated decoder – reusing it avoids re-building the decoder on
-# every call, which matters for repeat loads (e.g. force-refresh).
-_card_index_decoder: msgspec.json.Decoder[CardIndex] = msgspec.json.Decoder(CardIndex)
-
-
-def load_card_manager(data_dir: Path | str = CARD_DATA_DIR, force: bool = False) -> CardDataManager:
-    # Synchronous – call from a background thread. Downloads/updates card data if needed.
-    manager = CardDataManager(data_dir)
-    manager.ensure_latest(force=force)
-    return manager
-
-
-class CardDataManager:
-    def __init__(self, data_dir: Path | str = CARD_DATA_DIR):
-        self.data_dir = Path(data_dir)
-        self.data_dir.mkdir(parents=True, exist_ok=True)
-        # ``_v2`` filename invalidates pre-double-faced-aware caches.
-        self.index_path = self.data_dir / "atomic_cards_index_v2.json"
-        self.meta_path = self.data_dir / "atomic_cards_meta.json"
-        self._cards: list[CardEntry] | None = None
-        self._cards_by_name: dict[str, CardEntry] | None = None
-
-    def ensure_latest(self, force: bool = False) -> None:
-
-        remote_meta = self._fetch_remote_meta()
-        local_meta = self._load_meta_json() or {}
-        missing_index = not self.index_path.exists()
-        needs_refresh = force or missing_index
-        if not needs_refresh and remote_meta:
-            remote_size = remote_meta.get("content_length")
-            local_size = local_meta.get("content_length")
-            if remote_size and local_size != remote_size:
-                logger.warning(f"File size changed: local={local_size}, remote={remote_size}")
-                logger.warning("Forcing a refresh")
-                needs_refresh = True
-
-        if needs_refresh:
-            logger.warning("No atomic card index found or force refresh requested")
-            try:
-                if remote_meta:
-                    logger.info("Refreshing MTGJSON AtomicCards dataset")
-                else:
-                    logger.info("Fetching MTGJSON AtomicCards dataset (using headers for metadata)")
-                self._download_and_rebuild(remote_meta)
-            except Exception as exc:
-                if missing_index:
-                    raise RuntimeError(
-                        "Card data download failed and no cache is available"
-                    ) from exc
-                logger.warning(f"Failed to refresh MTGJSON data, using cache: {exc}")
-        self._load_index()
-
-    def search_cards(
-        self,
-        query: str = "",
-        format_filter: str | None = None,
-        type_filter: str | None = None,
-        color_identity: list[str] | None = None,
-        limit: int | None = None,
-    ) -> list[CardEntry]:
-        self._require_cards()
-        query = (query or "").strip().lower()
-        fmt = (format_filter or "").strip().lower()
-        type_filter = (type_filter or "").strip().lower()
-        color_identity = [c.upper() for c in (color_identity or [])]
-        results: list[CardEntry] = []
-        for card in self._cards or []:
-            name_lower = card.name_lower
-            type_line = (card.type_line or "").lower()
-            oracle_text = (card.oracle_text or "").lower()
-            if query:
-                haystacks = (
-                    name_lower,
-                    type_line,
-                    oracle_text,
-                )
-                if not any(query in h for h in haystacks if h):
-                    continue
-            if fmt and card.legalities.get(fmt) != "Legal":
-                continue
-            if type_filter and type_filter not in type_line:
-                continue
-            if color_identity:
-                identity = card.color_identity
-                if not all(c in identity for c in color_identity):
-                    continue
-            results.append(card)
-            if limit and len(results) >= limit:
-                break
-        return results
-
-    def get_card(self, name: str) -> CardEntry | None:
-        self._require_cards()
-        return (self._cards_by_name or {}).get(name.lower())
-
-    def available_formats(self) -> list[str]:
-        self._require_cards()
-        seen = set()
-        formats: list[str] = []
-        for card in self._cards or []:
-            for fmt, state in card.legalities.items():
-                if state != "Legal" or fmt in seen:
-                    continue
-                seen.add(fmt)
-                formats.append(fmt)
-        return sorted(formats)
-
-    @property
-    def is_loaded(self) -> bool:
-        return self._cards is not None
-
-    def _require_cards(self) -> None:
-        if self._cards is None:
-            raise RuntimeError("Card data not loaded; call ensure_latest first")
-
-    def _fetch_remote_meta(self) -> dict[str, Any] | None:
-        return self._fetch_dataset_headers()
-
-    def _fetch_dataset_headers(self) -> dict[str, Any] | None:
-        try:
-            resp = requests.head(
-                ATOMIC_DATA_URL,
-                impersonate="chrome",
-                timeout=ATOMIC_DATA_HEAD_TIMEOUT_SECONDS,
-            )
-            resp.raise_for_status()
-        except Exception as exc:
-            logger.warning(f"Failed to fetch MTGJSON dataset headers: {exc}")
-            return None
-        meta: dict[str, Any] = {}
-        headers = {k.lower(): v for k, v in resp.headers.items()}  # type: ignore[arg-type]
-        if "etag" in headers:
-            meta["etag"] = headers["etag"].strip('"')
-        if "last-modified" in headers:
-            meta["last_modified"] = headers["last-modified"]
-        if "content-length" in headers:
-            meta["content_length"] = headers["content-length"]
-        return meta or None
-
-    def _download_and_rebuild(self, remote_meta: dict[str, Any] | None) -> None:
-        resp = requests.get(
-            ATOMIC_DATA_URL,
-            impersonate="chrome",
-            timeout=ATOMIC_DATA_DOWNLOAD_TIMEOUT_SECONDS,
-        )
-        resp.raise_for_status()
-        content = resp.content
-        digest = hashlib.sha512(content).hexdigest()
-        with zipfile.ZipFile(io.BytesIO(content)) as zf:
-            with zf.open("AtomicCards.json") as source:
-                raw = json.load(source)
-        index = self._build_index(raw.get("data", {}))
-        atomic_write_json(self.index_path, index, ensure_ascii=False)
-        meta_to_store: dict[str, Any] = remote_meta.copy() if remote_meta else {}
-        meta_to_store.setdefault("sha512", digest)
-        headers = {k.lower(): v for k, v in resp.headers.items()}  # type: ignore[arg-type]
-        if "etag" in headers:
-            meta_to_store.setdefault("etag", headers["etag"].strip('"'))
-        if "last-modified" in headers:
-            meta_to_store.setdefault("last_modified", headers["last-modified"])
-        if "content-length" in headers:
-            meta_to_store.setdefault("content_length", headers["content-length"])
-        atomic_write_json(self.meta_path, meta_to_store, ensure_ascii=False)
-        self._cards = index["cards"]
-        self._cards_by_name = index["cards_by_name"]
-
-    def _load_index(self) -> None:
-        if not self.index_path.exists():
-            raise RuntimeError("Card data index missing or invalid")
-        try:
-            data = self.index_path.read_bytes()
-            card_index = _card_index_decoder.decode(data)
-        except (msgspec.DecodeError, OSError) as exc:
-            raise RuntimeError(f"Card data index missing or invalid: {exc}") from exc
-        self._cards = card_index.cards
-        self._cards_by_name = card_index.cards_by_name
-
-    def _load_meta_json(self) -> dict[str, Any] | None:
-        if not self.meta_path.exists():
-            return None
-        try:
-            return json.loads(self.meta_path.read_text(encoding="utf-8"))
-        except json.JSONDecodeError as exc:
-            logger.warning(f"Invalid JSON at {self.meta_path}: {exc}")
-            return None
-
-    def _build_index(self, atomic_cards: dict[str, list[dict[str, Any]]]) -> dict[str, Any]:
-        cards: dict[str, dict[str, Any]] = {}
-        alias_map: dict[str, dict[str, Any]] = {}
-        for variations in atomic_cards.values():
-            if not isinstance(variations, list) or not variations:
-                continue
-            printings = [
-                p for p in variations if not p.get("isToken") and p.get("layout") != "token"
-            ]
-            if not printings:
-                continue
-            canonical_name = (
-                printings[0].get("name") or printings[0].get("faceName") or ""
-            ).strip()
-            if not canonical_name:
-                continue
-            front_printing, back_printing = self._select_front_back(canonical_name, printings)
-            entry = self._simplify_printing(front_printing, canonical_name)
-            if back_printing is not None:
-                self._apply_back_face(entry, back_printing, canonical_name)
-            for printing in printings:
-                other = {k.lower(): v for k, v in (printing.get("legalities") or {}).items()}
-                entry["legalities"] = self._merge_legalities(entry.get("legalities"), other)
-            aliases = entry.setdefault("aliases", set())
-            for printing in printings:
-                aliases.update(self._collect_name_aliases(canonical_name, printing))
-            cards[canonical_name.lower()] = entry
-        card_list = sorted(cards.values(), key=lambda c: c["name_lower"])
-        for card in card_list:
-            alias_set = card.pop("aliases", set()) or set()
-            alias_set.add(card["name"])
-            cleaned_aliases = sorted({alias.strip() for alias in alias_set if alias})
-            card["aliases"] = cleaned_aliases
-            for alias in cleaned_aliases:
-                alias_map.setdefault(alias.lower(), card)
-        return {
-            "cards": card_list,
-            "cards_by_name": alias_map,
-        }
-
-    @staticmethod
-    def _select_front_back(
-        canonical_name: str,
-        printings: list[dict[str, Any]],
-    ) -> tuple[dict[str, Any], dict[str, Any] | None]:
-        """Pick which printing is the front face and which (if any) is the back.
-
-        Uses ``faceName`` matched against the part of ``canonical_name`` before
-        ``//``. Falls back to printings order when faceNames are missing.
-        """
-        if "//" not in canonical_name or len(printings) < 2:
-            return printings[0], None
-        front_part = canonical_name.split("//", 1)[0].strip().lower()
-        front_printing: dict[str, Any] | None = None
-        back_printing: dict[str, Any] | None = None
-        for printing in printings:
-            face_name = (printing.get("faceName") or "").strip().lower()
-            if face_name == front_part and front_printing is None:
-                front_printing = printing
-            elif back_printing is None:
-                back_printing = printing
-        if front_printing is None:
-            front_printing = printings[0]
-            if back_printing is None:
-                back_printing = printings[1]
-        return front_printing, back_printing
-
-    @staticmethod
-    def _apply_back_face(
-        entry: dict[str, Any],
-        printing: dict[str, Any],
-        canonical_name: str,
-    ) -> None:
-        face_name = (printing.get("faceName") or "").strip()
-        if not face_name and "//" in canonical_name:
-            parts = [p.strip() for p in canonical_name.split("//", 1)]
-            if len(parts) == 2:
-                face_name = parts[1]
-        entry["back_name"] = face_name or None
-        entry["back_mana_cost"] = printing.get("manaCost")
-        entry["back_type_line"] = printing.get("type")
-        entry["back_oracle_text"] = printing.get("text")
-        entry["back_power"] = printing.get("power")
-        entry["back_toughness"] = printing.get("toughness")
-        entry["back_loyalty"] = printing.get("loyalty")
-
-    def _simplify_printing(self, printing: dict[str, Any], canonical_name: str) -> dict[str, Any]:
-        legalities = printing.get("legalities") or {}
-        mana_value = printing.get("manaValue")
-        if isinstance(mana_value, str):
-            try:
-                mana_value = float(mana_value)
-            except ValueError:
-                pass
-        simplified = {
-            "name": canonical_name,
-            "name_lower": canonical_name.lower(),
-            "mana_cost": printing.get("manaCost"),
-            "mana_value": mana_value,
-            "type_line": printing.get("type"),
-            "oracle_text": printing.get("text"),
-            "power": printing.get("power"),
-            "toughness": printing.get("toughness"),
-            "loyalty": printing.get("loyalty"),
-            "colors": printing.get("colors") or [],
-            "color_identity": printing.get("colorIdentity") or [],
-            "legalities": {k.lower(): v for k, v in legalities.items()},
-            "aliases": set(),
-        }
-        return simplified
-
-    @staticmethod
-    def _collect_name_aliases(canonical_name: str, printing: dict[str, Any]) -> set[str]:
-        aliases: set[str] = set()
-        face_name = (printing.get("faceName") or "").strip()
-        if face_name:
-            aliases.add(face_name)
-        if canonical_name:
-            aliases.add(canonical_name)
-            if "//" in canonical_name:
-                for piece in canonical_name.split("//"):
-                    alias = piece.strip()
-                    if alias:
-                        aliases.add(alias)
-        return aliases
-
-    def _merge_legalities(
-        self,
-        base: dict[str, Any] | None,
-        incoming: dict[str, Any] | None,
-    ) -> dict[str, Any]:
-        merged: dict[str, Any] = {}
-        for source in (base or {}), (incoming or {}):
-            for fmt, state in source.items():
-                if state == "Legal":
-                    merged[fmt] = state
-        return merged
-
-
-__all__ = ["CardDataManager", "CardEntry", "load_card_manager"]
