+++ widgets/panels/card_inspector_panel/frame.py
@@ -12,7 +12,7 @@ from typing import Any
 import wx
 
 from services.image_service import CardImageRequest, get_cache
-from utils.card_data import CardDataManager
+from services.card_data_service import CardDataManager
 from utils.constants import (
     CARD_IMAGE_COST_MIN_HEIGHT,
     CARD_IMAGE_DISPLAY_HEIGHT,
