+++ widgets/panels/card_inspector_panel/handlers.py
@@ -11,7 +11,7 @@ import wx
 from loguru import logger
 
 from services.image_service import BULK_DATA_CACHE, CardImageRequest, get_card_image
-from utils.card_data import CardDataManager
+from services.card_data_service import CardDataManager
 from utils.constants import SUBDUED_TEXT, ZONE_TITLES
 
 if TYPE_CHECKING:
