+++ widgets/panels/card_inspector_panel/protocol.py
@@ -8,7 +8,7 @@ from typing import Any, Protocol
 import wx
 
 from services.image_service import CardImageRequest
-from utils.card_data import CardDataManager
+from services.card_data_service import CardDataManager
 from services.mana_icon_service import ManaIconFactory
 from widgets.panels.card_image_display import CardImageDisplay
 from widgets.panels.mana_rich_text_ctrl import ManaSymbolRichCtrl
