+++ widgets/panels/deck_stats_panel/protocol.py
@@ -8,7 +8,7 @@ import wx
 import wx.html2
 
 from services.deck_service import DeckService
-from utils.card_data import CardDataManager
+from services.card_data_service import CardDataManager
 
 
 class DeckStatsPanelProto(Protocol):
